Volunteen/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
captcha==0.5.0
colorama==0.4.6
crispy-bootstrap4==2023.1
Django==5.0.1
django-bootstrap4==24.1
django-bootstrap5==23.4
django-crispy-forms==2.1
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==6.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-two-factor-auth==1.15.2
phonenumbers==8.13.29
pillow==10.2.0
pypng==0.20220715.0
qrcode==7.4.2
setuptools==68.2.2
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
wheel==0.41.2
requests
django-qr-code==4.0.1
python-dotenv
psycopg2==2.9.10
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1